requirements.txt
absl-py==1.3.0
asgiref==3.6.0
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
Django==4.1.4
django-cors-headers==4.1.0
djangorestframework==3.14.0
drf-yasg==1.21.4
einops==0.2.0
filelock==3.13.1
flatbuffers==22.12.6
fonttools==4.44.3
fsspec==2023.10.0
gast==0.4.0
gdown==4.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
imageio==2.32.0
imgaug==0.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==14.0.6
lmdb==1.4.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==22.0
pandas==2.1.3
Pillow==9.3.0
prefetch-generator==1.0.1
protobuf==3.19.6
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.22.0
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.3
sympy==1.12
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
thop==0.1.1.post2209072238
tifffile==2023.9.26
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2022.7
ultralytics==8.0.210
uritemplate==4.1.1
urllib3==1.26.13
vietocr==0.3.12
Werkzeug==2.2.2
wrapt==1.14.1
